Eclipse Bug - Resolved with clean the project

+++ src/com/loiane/cursojava/aula56/GeraCpfCnpj.java
@@ -110,7 +110,7 @@ public class GeraCpfCnpj {
 				// converte o i-esimo caractere do CPF em um numero:
 				// por exemplo, transforma o caractere '0' no inteiro 0        
 				// (48 eh a posicao de '0' na tabela ASCII)        
-				num = (int) (CPF.charAt(i) - 48);
+				num = CPF.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso - 1;
 			}
@@ -125,7 +125,7 @@ public class GeraCpfCnpj {
 			sm = 0;
 			peso = 11;
 			for (i = 0; i < 10; i++) {
-				num = (int) (CPF.charAt(i) - 48);
+				num = CPF.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso - 1;
 			}
@@ -166,7 +166,7 @@ public class GeraCpfCnpj {
 				// converte o i-ésimo caractere do CNPJ em um número:
 				// por exemplo, transforma o caractere '0' no inteiro 0
 				// (48 eh a posição de '0' na tabela ASCII)
-				num = (int) (CNPJ.charAt(i) - 48);
+				num = CNPJ.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso + 1;
 				if (peso == 10)
@@ -183,7 +183,7 @@ public class GeraCpfCnpj {
 			sm = 0;
 			peso = 2;
 			for (i = 12; i >= 0; i--) {
-				num = (int) (CNPJ.charAt(i) - 48);
+				num = CNPJ.charAt(i) - 48;
 				sm = sm + (num * peso);
 				peso = peso + 1;
 				if (peso == 10)
